Functions and Images
I have created the basic functions needed for my project, as well as found the images needed to demonstrate the CT stretches. I have also imported "time" to use in order to create the scheduling system for the app. The functions do not yet work, as the time related variables haven't been set to any values.