Created Current Time Variable and IF/ELIF Statements
In this update, I have created the "current_time" variable that will find and print the current time of day. I have also created the IF/ELIF statements that will call the functions I've created. There are no syntax errors in the code and it will allow an input at the beginning but upon input the code does not work.