Added Music, Video Demonstrations, New Functions
I've added music to play during the stretches, a warm-up tutorial video as well as two more tutorial videos for the flexor and extensor stretches. I added VLC and OS imports to play the  music and videos repectively. This update also includes the functions needed for the videos and music to work.