Added new exercise, video skip, PowerPoint, time list
In this update, I have added a new exercise function: the Tendon Glide. This makes 3, and I plan to add a 4th by the end of the day today. I have also created versions of my functions that skip the practice videos for returning users. I've also added my PowerPoint describing the purpose of MyHandTracker. Last but not least, I am now prompting the user to enter 4 different times in the time_hourmin string and am using .split() to turn the string into a list. From there, I plan to have HandTracker go through the list to check the time to see if it's time for a stretch session.